+++ src/Components/Input.php
@@ -22,7 +22,7 @@ class Input extends FormControl
         string $type = 'text',
         string $name = null,
         string $id = null,
-        string|array $value = [],
+        string|array $value = null,
         string $current = null,
         string $default = null,
         bool $invalid = false,