+++ src/Components/Input/Time.php
@@ -9,7 +9,7 @@ class Time extends Input
     public function __construct(
         string $name = null,
         string $id = null,
-        string|array $value = [],
+        string|array $value = null,
         string $current = null,
         string $default = null,
         bool $invalid = false,