+++ CHANGELOG.md
@@ -6,9 +6,11 @@ All notable changes to [Sikessem's UI kit](https://github.com/sikessem/ui) will 
 
 **Full Changelog:** [v0.5.4...v0.6.0](https://github.com/sikessem/ui/compare/v0.5.4...v0.6.0)
 
+- [0.x] Fix button text
+- [0.x] Add text element
+
 ## [v0.6.0](https://github.com/sikessem/ui/releases/tag/v0.6.0) - 2023-08-07
 
-- [0.x] Add text element
 - [0.x] Add select component
 - [0.x] Improve entry component
 - [0.x] Add textarea component